refactor make_sysex sans malloc

+++ src/c/lib/roland.h
@@ -9,8 +9,8 @@ typedef struct roland_sysex {
 	u8 data[];
 } RolandSysex;
 
-size_t         make_receive_sysex(u8 *, Addr , Addr );
-size_t         make_send_sysex(u8 *, Addr, u8 *, size_t );
-size_t         make_sysex(u8 *, u8, size_t);
-RolandSysex *  parse_sysex(u8 *, size_t );
+int         make_receive_sysex(u8 *, Addr , Addr );
+int         make_send_sysex(u8 *, Addr, u8 *, int );
+int         make_sysex(u8 *, u8, int);
+RolandSysex *  parse_sysex(u8 *, int );
